make build: fix for case of configuration file in top-level folder
When the configuration file (distortosConfiguration.mk) is located in
the top-level folder of the project the script that generates the linker
script fails with following error message:

./source/chip/STMicroelectronics/STM32F4/STM32F4.ld.sh
"distortosConfiguration.mk" > "output/STM32F407VGT.ld"
./source/chip/STMicroelectronics/STM32F4/STM32F4.ld.sh: line 28: .:
distortosConfiguration.mk: file not found

This is because shell doesn't like sourcing files without path (with no
slashes). Update the rule in Rules.mk to always include the directory
component ("./" in this particular case) to fix this problem.