ThreadCommon::start(): move check of state to Scheduler::add()
The thread cannot be started again once it was started. The check of
thread's state in ThreadCommon::start() was vulnerable to a race
condition - the check may pass in one thread but context is switched
before call to Scheduler::add() and ThreadCommon::start() is called for
the same object from another thread. By moving the check to
Scheduler::add() - which is protected with critical zone - possibility
of such race condition is eliminated.