add Thread::detach()
This function is present only if CONFIG_THREAD_DETACH_ENABLE option is
enabled.

All required implementations are provided:
1. UndetachableThread::detach() - always fails with ENOTSUP error code;
2. DynamicThread::detach() - releases detachable thread managed by
unique_ptr and calls DynamicThreadBase::detach();
3. DynamicThreadBase::detach() - resets the pointer to owner
DynamicThread and calls DynamicThread::detach();

The recursive calls to detach() (DynamicThread::detach() ->
DynamicThreadBase::detach() and DynamicThreadBase::detach() ->
DynamicThread::detach()) are handled properly - second call to the same
function fails with an error and recursive calls handle this error,
treating it as success (because it is actually expected in this case).