Add regenerateKconfig.sh
This script will locate all filenames beginning with "Kconfig". In the
found files it will try to locate all blocks which have the following
pattern:

<regenerated-block>

It will then replace whole <regenerated-block> with lines in the
following form:

source "$DISTORTOS_PATH/<found-file>"

In these lines <found-file> is the filename matching the <pattern> of
regenerated block.

This script can be also used from "outside" of distortos source tree
(e.g. in a project based on "subfolder" template). In this case it
should be executed with additional argument - path to distortos folder.