Add generateKconfig.sh script
This script should be called with one argument - path to folder in which
the files will be generated. Default value for this argument is "output"
(default value of $(OUTPUT) variable in the build scripts).

The script searches for all files matching "Kconfig*" pattern. Each file
is scanned for pattern in the following form

"### don't remove this line! <filename>"

For each pattern a file named <filename> is generated in the selected
folder. This generated file includes ("sources") all files named
<filename> (except itself) that were found in the project.

This generated file can be included by "normal" Kconfig scripts and it
will act as a "proxy" file for other files that can be freely
added/removed. This is useful for adding custom boards, user application
configuration or for removing unused parts of distortos (like unused
chips, architectures, boards, test application, ...).

The script calls "find" with full path as "/usr/bin/find" in an attempt
to simplify requirements for Windows. In Windows there's a "find.exe"
file in "c:/Windows/System32", which conflicts with "find.exe" from
MSYS2. One option is to require MSYS2 folder to be at the very beginning
of system's PATH environment variable, the other option is to use full
"/usr/bin/..." path. Both options are bad, but the latter seemed like a
smaller inconvenience for users.