Fix dynamic memory checks in test cases for "official" ARM toolchains
New versions of "official" ARM toolchain (5.x) allocate some dynamic
memory somewhere in __libc_init_array(). Test cases must not assume that
"all" dynamic memory is deallocated at the end of test case. It's enough
to verify that the amount of allocated memory is the same as when the
test case was started.