Limit write transfers in SerialPort to half of buffer size
This makes SerialPort::startWriteWrapper() more consistent with
SerialPort::startReadWrapper(). Limiting the size of single write
transfer to half the size of circular buffer is a potential optimization
for cases where write() call with small amount of data is made while big
transfer is already in progress. If this transfer is already past half
the size of buffer, then the other half will be available for writing.
The thread which does the call to write() will not need to restart the
transfer that is in progress and the new data can be written to the
circular buffer immediately.