Improve algorithm of SerialPort::read()
If blocking is required, read transfers are operating directly on the
output buffer. This way there's no need to copy the data from circular
buffer to output buffer and only one blocking operation is needed. 

Additionally this nicely handles the case when semaphore wait is
interrupted but the thread that's in read() is not immediately executed
(due to some higher priority thread running) - the data will be read to
the output buffer as long as there is any space left and as long as the
thread is not resumed. The same will be true when timeouts will be
implemented.