Improve algorithm of SerialPort::write()
If blocking is required, write transfers are operating directly on the
input buffer. This way there's no need to copy the data from input
buffer to circular buffer and only one blocking operation is needed. 

Additionally this nicely handles the case when semaphore wait is
interrupted but the thread that's in write() is not immediately executed
(due to some higher priority thread running) - the data will be written
from the input buffer as long as there is any data left and as long as
the thread is not resumed. The same will be true when timeouts will be
implemented.