Fix bug in periodic software timers
During execution of SoftwareTimerControlBlock::run() the timer is
technically "stopped" (it's not linked), but in case of periodic timers
this is wrong. This is a problem because
SoftwareTimerControlBlock::stopInternal() does nothing if
SoftwareTimerControlBlock::isRunning() returns false.

Improve SoftwareTimerControlBlock::isRunning() to report real state of
timer, even for periodic timers. Then change the restart test in
SoftwareTimerControlBlock::run() - if the timer was restarted within
timer's function, then its node will be linked, no need to use
SoftwareTimerControlBlock::isRunning().