Don't needlessly change SPE bit in CR1 register of STM32F4's SPI
Don't enable/disable SPI peripheral at the beginning and end of each
transfer, as this causes all SPI lines to to into HiZ state. This may
cause single clock pulses to appear on SCLK line (for example if CPOL is
0 and SPI lines have pull-ups), messing the operation of slave device if
it is still selected.

Because the peripheral is now enabled as long as the driver is started,
it's necessary to improve ChipSpiMasterLowLevel::configure(). Value of
DFF bit (which determines word length) must be changed only when SPI
peripheral is disabled.