Don't use "realpath" in generateKconfig.sh
"realpath" is not a standard solution - the "--relative-base=..." switch
is not supported on all platforms. Normalize path with "cd" and "pwd",
which seems slightly more complicated, but is more portable.

There is no point in having the argument optional, so now the script
exits with an error if no argument is given.

Fixes #22