Add implementations of bit-band macros for STM32's FLASH registers
Due to inconsistency of ST headers, "normal" STM32_BITBAND() and
STM32_BITBAND_ADDRESS() macros cannot be used with FLASH registers. This
is because for FLASH the macro FLASH_BASE is the base address of flash
memory (0x8000000), not the base address of FLASH registers. The correct
macro in this case is FLASH_R_BASE, used by new STM32_BITBAND_FLASH()
and STM32_BITBAND_FLASH_ADDRESS().