Use dict.items() instead of dict.iteritems()
In Python 3 dict.iteritems() is not present, but dict.items() is present
in both Python 2 and 3. Functionally dict.items() of Python 3 is
equivalent of dict.iteritems() from Python 2. In Python 2 dict.items()
may cause high memory usage, as it actually returns a list, not a
iterator/view. For distortos this shouldn't be a problem, as the
dictionaries are not expected to be really big.