Add pydts.py script
This script uses PLY to parse *.dts files into a Python dictionary.
Internal grammar was greatly inspired by original dtc. The format of
arguments is also similar.

As this is a proof-of-concept for now, only limited set of devicetree
syntax is supported. Following features are currently NOT handled:
- path references;
- bytestrings;
- any integer arithmetic other than addition and multiplication;
- labels for anything else than a node;
- multiple labels for single node;
- memory reservations;
- any non-standard extensions;
- /include/ statements - #include statements should be used exclusively;
- C or C++ comments - handled by C preprocessor;

The features which are currently handled are enough to parse all STM32
*.dts files from Linux or from Zephyr.

The script relies on *C preprocessor* to resolve #define macros, handle
#include statements and remove C/C++ comments. Most likely before you
feed the file into this script, you should first do sth like this:

$ cpp -nostdinc -undef -E -x assembler-with-cpp in.dts > out.dts

Linemarkers generated by current `cpp` versions are supported.