Use real object for system's Scheduler
GCC 4.x did not implement constexpr constructors fully, so in some cases
the object was placed in .bss and initialized during run-time. This was
fixed in GCC 5 - constexpr constructor of Scheduler really causes the
object to be placed in .data section, without any run-time
initialization. As GCC 4.x is no longer supported by distortos,
"placement new" initialization of Scheduler in aligned_storage can be
removed.