Fix (?) SoftwareTimerControlBlock::isRunning() when LTO is enabled
Add memory barrier to SoftwareTimerControlBlock::isRunning(), which
fixes the cases where this function is executed in a loop (possibly as a
loop termination condition) or multiple times in a row. Without the
memory barrier when LTO is enabled, the function is called only once and
the loop gets converted to an infinite version.