Add support for arbitrary pattern to generateKconfig.sh
This script will scan the Kconfig files looking for two kinds of special
lines which will then be used to generate output "grouping" files. The
first kind of pattern is the "basic" one, where the name of output
"grouping" file is identical to the names of "grouped" files. An example
of this pattern looks like this:

source "$OUTPUT/Kconfig-applicationOptions"

In this case generated file will be named "Kconfig-applicationOptions",
and it will contain list of all found files which were also named
"Kconfig-applicationOptions".

The new variant of this pattern allows the search pattern to be
completely arbitrary, containing also shell globs. An example would look
like that:

source "$OUTPUT/Kconfig-STM32F4" # pattern: Kconfig-STM32F4*

In this case generated file will be named "Kconfig-STM32F4", and it will
contain list of all found files which matched "Kconfig-STM32F4*" pattern
(for example "Kconfig-STM32F401RE", "Kconfig-STM32F407VG" and so on).

The pattern is passed verbatim to `find`, so any feature supported by
this tool may be used.