Zero-initialize objects before each test in C-API-Semaphore-unit-test-1
Depending on host architecture, the objects may have padding at the end,
which will _not_ be set by constructors. As the memory is zeroed during
the test, some comparisons may fail unexpectedly, as the padding bytes
at the end will be "random" in the "constructed" buffer and zeroed in
the "real" object. Solve this by zeroing the objects at the beginning of
each test.