Fix PrettyPrinters according to PEP 257
According to PEP 257 (https://www.python.org/dev/peps/pep-0257/):
- Triple quotes are used even though the string fits on one line.
This makes it easy to later expand it.
- There's no blank line either before or after the docstring (Used more
lenient version in classes).
- The docstring is a phrase ending in a period. It prescribes the
function or method's effect as a command ("Do this", "Return that"), not
as a description; e.g. don't write "Returns the pathname ..." (Used more
lenient version in classes).