Split text and vectors in generated linker scripts
This change allows many different customizations for the project. By
altering the address & size of main flash and adding another region just
for the vectors (located at the beginning) - both done via board's YAML
file - you can use the script in following scenarios:

1. EEPROM emulation located on the second and third page of flash

This can be useful for chips such as STM32F4, where first four pages of
flash are 16 kB each, while the last pages (traditionally used for
EEPROM emulation) are exteremely big (128 kB) and may not be "paired"
(for example in some chips the second to last page is 64 kB, while the
the last one is 128 kB). For that scenario change the address and size
of main flash to start at the very beginning of fourth page and add new
region which covers whole first page.

2. Firmware with embedded size and checksum

This can be useful when your project uses custom bootloader, which
verifies whether the firmware is valid. For this scenario you usually
need the size of whole firmware at the first available place - which is
just after the vectors - and the checksum as the last entry in the
flash (note that embedding the checksum requires a two step linking
process). In this case modify the address and size of main flash region
to start after the vectors and embedded size of firmware (so
size of vectors + 4 bytes) and add new region which is at the beginning
of flash. In your custom additional linker script (linking will use both
scripts concatenated together) put the size of firmware into the
same region as ".text.vectors" and add new section for the checksum in
the same region as ".text".