Change "total..." to "...Count" when refering to number of outputs
This way the name of the object (`dictionaryKey`) doesn't have to be
modified to start with uppercase letter. Jinja's `capitalize` filter
doesn't work very well for multiword camelCase identifiers, so if
`dictionaryKey`  happened to be `userOutputs`, generated identifier
would be `totalUseroutputs`. After the change no modification of casing
is necessary, as the generated identifier is `userOutputsCount`.

Regenerate all boards:
$ find source/board/ -name '*.yaml' -print0 | xargs -0 -I{} -t make
board CONFIG_FILE={}