Fix CMake dependencies of distortos::distortos target
Explicitly use `distortos` instead of `-ldistortos` in
target_link_libraries() associated with distortos::distortos. This way
proper link dependency between the application and the static library is
created. Without this fix any change in distortos source code would
cause the library to be rebuilt, but the application would not be linked
again.