Send dummy byte before every command instead of after selecting
For most of the cases sending dummy byte in SelectGuard's constructor is
the same as sending it before every command in writeCmd(). However for
ACMDx there is a difference. Sending ACMDx technically means sending two
separate commands. With dummy byte in the constructor there will be no
dummy byte between CMD55 and the actual ACMDx. For some cards this makes
no difference, however older cards just go crazy when this particular
dummy byte is missing and respond with completely invalid format (0xe0
where R1 response is expected) or respond that ACMD41 is not a supported
command (0x05 R1 response).

Tested with Sandisk 2 GB SDSC and Kingston 2 GB SDSC.