Make generated distortosConfiguration.cmake more portable
Due to the way CMake works, these files have hardcoded relative path to
the toolchain file (CMAKE_TOOLCHAIN_FILE). If the relative path from the
build folder to the toolchain file changes, the file is unusable (unless
you edit it by hand). This may happen if you would like to use the
configuration created in the distortos project as a base of your own
project, based on the distortosTemplateSubfolder organization.

Add an option to work-around this issue:
- path to CMAKE_TOOLCHAIN_FILE in distortosConfiguration.cmake is always
taken relative to distortos folder, not the project top-level folder,
- path to CMAKE_TOOLCHAIN_FILE in distortosConfiguration.cmake is
prefixed with additional variable, which by default is set to relative
path from build folder to distortos folder, but may be overriden by
DISTORTOS_PATH environment variable.

With these changes in place, files from one project can be used in
another project in following way:

$ DISTORTOS_PATH=../distortos/ cmake -C
../distortos/configurations/ST_STM32F4DISCOVERY/test/distortosConfiguration.cmake
.. -GNinja