Require DISTORTOS_BOARD_VERSION from generated board to be exactly 0
This is a safety test for cases where code in distortos is changed in a
way that requires board to be regenerated. For example currently core
and chip vectors were removed and replaced with generated sources in the
board. If user updates distortos, but does not regenerate board, the
application will build and link with no issues, but it will completely
fail to work because no vectors are present in the executable.