Fix some CMake conditions in `distortosSetConfiguration()` handling
When using `cmake_parse_arguments()` checking for keywords should be
done with `if(DEFINED <prefix>_<keyword>)`. Using just
`if(<prefix>_<keyword>)` will result in FALSE when the keyword is
defined but its value is 0.

This is a problem in case of `distortosSetConfiguration(INTEGER ...)`,
where `MIN` or `MAX` may be set to zero.