Initially hide all cache variables instead of removing inactive ones
The main benefit of this approach is that hiding is done at the very
beginning, while removal has to be done at the very end of configuration
phase. When distortos is used as a submodule of a project, then there is
no control over the "end", while it is easy to know the beginning.

Inactive variables will never be removed from the cache, but this is
just a very minor issue. As a benefit the removed variables are not
reset to their defaults when they reappear later.

Regenerate all configurations (they are no longer sorted
alphabetically):
$ ./scripts/forAllConfigurations.sh . cp distortosConfiguration.cmake
../\${configuration}