Modify mechanism of generating distortosConfiguration.cmake
Currently this file is generated "on-the-fly". This solves the problem
that any additional variable added after the generation would not be
placed in the header.

distortosAppendToSavedConfiguration() detects when it is used during
compiler identification and checking and does nothing in that case. It
also handles deferred writes, which may happen when configuration
variable is added before distortos project is defined (for example when
distortos is just a submodule and toolchain file adds something).

Regenerate all configurations:
$ ./scripts/forAllConfigurations.sh . cp distortosConfiguration.cmake
../\${configuration}