Don't use PROVIDE() for restricted identifiers in linker scripts
All the symbols in linker script are restricted (prefixed with double
underscore), so there is actually no need to use PROVIDE() to hide them
in case application redefines them. Generating them in all cases (which
is what removal of PROVIDE() achieves) will allow the linker script to
be simplified and also used in some special cases in bare-metal
applications.

Regenerate all boards:
$ find source/board/ -name '*.yaml' -print0 | xargs -0 -I{} -t
./scripts/generateBoard.py {}