Improve error codes returned by VirtualFileSystem::makeDirectory()
Some cases where ENOENT was returned should actually fail with different
code when the path is just a single component: either EROFS (if it was
not matched to the mount point) or EEXIST (otherwise). At this level
ENOENT should be returned only if the path has more than one component,
but the first one was not matched to the existing mount point.