Really fix warning in estd::extractBitField() about shift count overflow
It turns out that recent fix was only partial, as it doesn't change
anything for GCC 6. Fix it now by inverting the shift logic. Instead of
potentially causing shift overflow in left shift, use right shift in a
way that does not allow for any overflow - max shift is now limited to
"1 bit less than width of type".