Fix overrun handling in STM32-USARTv1's ChipUartLowLevel driver
It is possible for overrun to be detected between reads of SR and DR. In
this case old code would get stuck in an endless interrupt - RXNEIE is
enabled, ORE flag is set so interrupt is pending, but the interrupt
handler does not handle this condition, because RXNE flag is not set.

Fix this by slightly reworking the loop and adding the explicit check
for ORE flag.

Note that similar handling is most likely not needed for other error
flags, as they either don't generate interrupt or they have a separate
"interrupt enabled" switch.