Make test configuration for ST NUCLEO-U575ZI-Q board
Because there is no standard clock configuration available for this chip
family yet and because this chip uses a very slow clock after reset (4
MHz), the test configurtion has to use reduced tick frequency of 100 Hz.
It will fail during the tests with default tick frequency of 1000 Hz,
because some test case steps are expected to complete within single
tick.